langgraph-agent-observability/requirements.txt
# Langgraph Requirements for Observability
boto3
botocore
langchain[aws]
langgraph
langsmith[otel]
ddgs
langchain-community
opentelemetry-instrumentation-langchain
aws-opentelemetry-distro~=0.12.1
python-dotenv
opentelemetry-api